Fix for Ou dropping migration - OUs don't necessarily exist in all databases

+++ db/migrate/20140129223722_ensure_ou_table_is_removed.rb
@@ -1,5 +1,7 @@
 class EnsureOuTableIsRemoved < ActiveRecord::Migration
   def change
-    drop_table :ous
+    if defined? Ou
+      drop_table :ous
+    end
   end
 end
